Revert "Update WallPaper.cpp"
This reverts commit a6563050ad63b46e8dccae557fbb645d6791f3c9.

+++ Source/WallPaper/Private/WallPaper.cpp
@@ -113,8 +113,13 @@ void FWallPaperModule::ShutdownModule()
 	WallpaperCommands::Unregister();
 	StyleSettings->ResetStyleColor();
 	//Clear Cache
-
-	FString TargetFilePath = FPaths::ProjectSavedDir()/"Cache";
+	FString PluginsPath = FPaths::ProjectPluginsDir()/"Wallpaper";
+	if(!IFileManager::Get().DirectoryExists(*PluginsPath))
+    	{
+    		PluginsPath = FPaths::EnginePluginsDir()/"Wallpaper";
+    	}
+    		
+    	FString TargetFilePath = PluginsPath/"Content/Cache";
     	if(IFileManager::Get().DirectoryExists(*TargetFilePath))
     	{
     		IFileManager::Get().DeleteDirectory(*(TargetFilePath),false,true);
@@ -614,7 +619,7 @@ void FWallPaperModule::ImportPicTheme()
 		PluginsPath = FPaths::EnginePluginsDir()/"Marketplace/Wallpaper";
 	}
 		
-	FString TargetFilePath = FPaths::ProjectSavedDir()/"Cache";
+	FString TargetFilePath = PluginsPath/"Content/Cache";
 	//clear file
 	if(IFileManager::Get().DirectoryExists(*TargetFilePath))
 	{
